refactor: split admin panel UI out of admin_commands (#104)
Keep AdminCommands focused on slash command orchestration while moving panel views, buttons, and modals into a dedicated package. This reduces churn in the cog without constraining a later refactor of user command flows.
