+++ typer_bot/commands/admin_commands.py
@@ -3,12 +3,12 @@
 from __future__ import annotations
 
 import logging
-from dataclasses import dataclass
 
 import discord
 from discord import app_commands
 from discord.ext import commands
 
+from typer_bot.commands.admin_panel import AdminPanelHomeView, DeleteConfirmView
 from typer_bot.database import Database
 from typer_bot.handlers import FixtureCreationHandler, ResultsEntryHandler
 from typer_bot.services import AdminService
@@ -21,7 +21,6 @@ from typer_bot.utils.db_backup import cleanup_old_backups, create_backup
 _calculate_cooldowns: dict[str, float] = {}
 CALCULATE_COOLDOWN = 30.0
 COOLDOWN_EXPIRY_HOURS = 1
-MAX_SELECT_OPTIONS = 25
 
 logger = logging.getLogger(__name__)
 
@@ -52,37 +51,6 @@ def admin_only():
         return True
 
     return app_commands.check(predicate)
-
-
-def _fixture_select_label(fixture: dict) -> str:
-    status = fixture["status"].upper()
-    return f"Week {fixture['week_number']} [{status}]"
-
-
-def _fixture_select_description(fixture: dict) -> str:
-    games_count = len(fixture["games"])
-    return f"{games_count} matches"
-
-
-def _format_prediction_line(index: int, game: str, prediction: str) -> str:
-    return f"{index}. {game} {prediction}"
-
-
-def _prediction_status_text(prediction: dict) -> str:
-    if not prediction["is_late"]:
-        return "on time"
-    if prediction["late_penalty_waived"]:
-        return "late, waiver active"
-    return "late, penalty active"
-
-
-@dataclass(slots=True)
-class PanelSelectionState:
-    """Shared selection state for admin panel flows."""
-
-    fixture_id: int | None = None
-    user_id: str | None = None
-    status_message: str = ""
 
 
 class AdminCommands(commands.Cog):
@@ -391,600 +359,6 @@ class AdminCommands(commands.Cog):
             f"{preview}\n\nMention users in this post?",
             view=view,
             ephemeral=True,
-        )
-
-
-class OwnerRestrictedView(discord.ui.View):
-    """View base class that restricts interactions to one admin."""
-
-    def __init__(self, admin_cog: AdminCommands, owner_user_id: str, timeout: float = 180):
-        super().__init__(timeout=timeout)
-        self.admin_cog = admin_cog
-        self.owner_user_id = owner_user_id
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if str(interaction.user.id) != self.owner_user_id:
-            await interaction.response.send_message(
-                "You don't have permission to do this!", ephemeral=True
-            )
-            return False
-        if not is_admin(interaction):
-            await interaction.response.send_message(
-                "You no longer have permission to use admin commands.", ephemeral=True
-            )
-            return False
-        return True
-
-
-class AdminPanelHomeView(OwnerRestrictedView):
-    @discord.ui.button(label="Fixtures", style=discord.ButtonStyle.secondary)
-    async def fixtures(self, interaction: discord.Interaction, _button: discord.ui.Button):
-        view = FixturesPanelView(self.admin_cog, self.owner_user_id)
-        await view.load_fixture_options()
-        await interaction.response.edit_message(content=view.render_content(), view=view)
-
-    @discord.ui.button(label="Predictions", style=discord.ButtonStyle.primary)
-    async def predictions(self, interaction: discord.Interaction, _button: discord.ui.Button):
-        view = PredictionsPanelView(self.admin_cog, self.owner_user_id)
-        await view.load_fixture_options()
-        await interaction.response.edit_message(content=view.render_content(), view=view)
-
-    @discord.ui.button(label="Results", style=discord.ButtonStyle.success)
-    async def results(self, interaction: discord.Interaction, _button: discord.ui.Button):
-        view = ResultsPanelView(self.admin_cog, self.owner_user_id)
-        await view.load_fixture_options()
-        await interaction.response.edit_message(content=view.render_content(), view=view)
-
-
-class FixturesPanelView(OwnerRestrictedView):
-    """Panel for fixture deletion and workflow guidance."""
-
-    def __init__(self, admin_cog: AdminCommands, owner_user_id: str):
-        super().__init__(admin_cog, owner_user_id)
-        self.selection = PanelSelectionState()
-        self.fixture_select = FixtureSelect(self, open_only=True)
-        self._refresh_items()
-
-    def _refresh_items(self) -> None:
-        self.clear_items()
-        self.add_item(self.fixture_select)
-        self.add_item(FixturesDeleteButton(self))
-        self.add_item(BackButton(self))
-
-    async def load_fixture_options(self) -> None:
-        fixtures = await self.admin_cog.db.get_open_fixtures()
-        self.fixture_select.update_options(fixtures)
-
-    def render_content(self) -> str:
-        status = (
-            self.selection.status_message
-            or "Select an open fixture to delete, or use `/admin fixture create` for new fixtures."
-        )
-        return "**Admin Panel - Fixtures**\n" + status
-
-
-class PredictionsPanelView(OwnerRestrictedView):
-    """Panel for prediction lookup and override actions."""
-
-    def __init__(self, admin_cog: AdminCommands, owner_user_id: str):
-        super().__init__(admin_cog, owner_user_id)
-        self.selection = PanelSelectionState()
-        self.fixture_select = FixtureSelect(self)
-        self.user_select = PredictionUserSelect(self)
-        self.user_select.update_options([])
-        self._refresh_items()
-
-    def _refresh_items(self) -> None:
-        self.clear_items()
-        self.add_item(self.fixture_select)
-        self.add_item(self.user_select)
-        self.add_item(ViewPredictionsButton(self))
-        self.add_item(ReplacePredictionButton(self))
-        self.add_item(ToggleWaiverButton(self))
-        self.add_item(BackButton(self))
-
-    async def load_fixture_options(self) -> None:
-        fixtures = await self.admin_cog.service.get_recent_fixtures(MAX_SELECT_OPTIONS)
-        self.fixture_select.update_options(fixtures)
-
-    async def load_user_options(self) -> None:
-        if self.selection.fixture_id is None:
-            self.user_select.update_options([])
-            return
-        predictions = await self.admin_cog.db.get_all_predictions(self.selection.fixture_id)
-        self.user_select.update_options(predictions)
-
-    def render_content(self) -> str:
-        status = self.selection.status_message or (
-            "Select a fixture, then pick a user to view or override a stored prediction."
-        )
-        return "**Admin Panel - Predictions**\n" + status
-
-
-class ResultsPanelView(OwnerRestrictedView):
-    """Panel for result correction workflows."""
-
-    def __init__(self, admin_cog: AdminCommands, owner_user_id: str):
-        super().__init__(admin_cog, owner_user_id)
-        self.selection = PanelSelectionState()
-        self.fixture_select = FixtureSelect(self)
-        self._refresh_items()
-
-    def _refresh_items(self) -> None:
-        self.clear_items()
-        self.add_item(self.fixture_select)
-        self.add_item(ViewResultsButton(self))
-        self.add_item(CorrectResultsButton(self))
-        self.add_item(BackButton(self))
-
-    async def load_fixture_options(self) -> None:
-        fixtures = await self.admin_cog.service.get_recent_fixtures(MAX_SELECT_OPTIONS)
-        self.fixture_select.update_options(fixtures)
-
-    def render_content(self) -> str:
-        status = (
-            self.selection.status_message or "Select a fixture to inspect or correct saved results."
-        )
-        return "**Admin Panel - Results**\n" + status
-
-
-class FixtureSelect(discord.ui.Select):
-    def __init__(
-        self,
-        parent_view: FixturesPanelView | PredictionsPanelView | ResultsPanelView,
-        open_only: bool = False,
-    ):
-        self.parent_view = parent_view
-        self.open_only = open_only
-        super().__init__(
-            placeholder="Select fixture",
-            min_values=1,
-            max_values=1,
-            disabled=True,
-        )
-
-    def update_options(self, fixtures: list[dict]) -> None:
-        if self.open_only:
-            fixtures = [fixture for fixture in fixtures if fixture["status"] == "open"]
-
-        if not fixtures:
-            self.options = [discord.SelectOption(label="No fixtures available", value="none")]
-            self.disabled = True
-            return
-
-        self.options = [
-            discord.SelectOption(
-                label=_fixture_select_label(fixture),
-                value=str(fixture["id"]),
-                description=_fixture_select_description(fixture),
-            )
-            for fixture in fixtures[:MAX_SELECT_OPTIONS]
-        ]
-        self.disabled = False
-
-    async def callback(self, interaction: discord.Interaction):
-        if self.values[0] == "none":
-            await interaction.response.send_message("No fixtures available.", ephemeral=True)
-            return
-
-        fixture_id = int(self.values[0])
-        self.parent_view.selection.fixture_id = fixture_id
-        self.parent_view.selection.user_id = None
-
-        fixture = await self.parent_view.admin_cog.db.get_fixture_by_id(fixture_id)
-        if fixture is None:
-            self.parent_view.selection.status_message = "Fixture no longer exists."
-        else:
-            self.parent_view.selection.status_message = (
-                f"Selected week {fixture['week_number']} ({fixture['status']})."
-            )
-
-        if isinstance(self.parent_view, PredictionsPanelView):
-            await self.parent_view.load_user_options()
-
-        await interaction.response.edit_message(
-            content=self.parent_view.render_content(),
-            view=self.parent_view,
-        )
-
-
-class PredictionUserSelect(discord.ui.Select):
-    """Select a user who already has a prediction for the chosen fixture."""
-
-    def __init__(self, parent_view: PredictionsPanelView):
-        self.parent_view = parent_view
-        super().__init__(
-            placeholder="Select user",
-            min_values=1,
-            max_values=1,
-            disabled=True,
-        )
-
-    def update_options(self, predictions: list[dict]) -> None:
-        if not predictions:
-            self.options = [discord.SelectOption(label="No predictions available", value="none")]
-            self.disabled = True
-            return
-
-        ordered = sorted(predictions, key=lambda prediction: prediction["user_name"].lower())
-        self.options = [
-            discord.SelectOption(
-                label=prediction["user_name"][:100],
-                value=prediction["user_id"],
-                description=_prediction_status_text(prediction)[:100],
-            )
-            for prediction in ordered[:MAX_SELECT_OPTIONS]
-        ]
-        self.disabled = False
-
-    async def callback(self, interaction: discord.Interaction):
-        if self.values[0] == "none":
-            await interaction.response.send_message("No predictions available.", ephemeral=True)
-            return
-
-        self.parent_view.selection.user_id = self.values[0]
-        fixture_id = self.parent_view.selection.fixture_id
-        if fixture_id is None:
-            await interaction.response.send_message("Select a fixture first.", ephemeral=True)
-            return
-
-        prediction = await self.parent_view.admin_cog.db.get_prediction(
-            fixture_id,
-            self.values[0],
-        )
-        if prediction is None:
-            self.parent_view.selection.status_message = "Prediction no longer exists."
-        else:
-            self.parent_view.selection.status_message = (
-                f"Selected {prediction['user_name']} ({_prediction_status_text(prediction)})."
-            )
-
-        await interaction.response.edit_message(
-            content=self.parent_view.render_content(),
-            view=self.parent_view,
-        )
-
-
-class BackButton(discord.ui.Button):
-    def __init__(self, parent_view: OwnerRestrictedView):
-        self.parent_view = parent_view
-        super().__init__(label="Back", style=discord.ButtonStyle.secondary)
-
-    async def callback(self, interaction: discord.Interaction):
-        view = AdminPanelHomeView(self.parent_view.admin_cog, self.parent_view.owner_user_id)
-        await interaction.response.edit_message(
-            content="**Admin Panel**\nChoose the workflow you want to manage.",
-            view=view,
-        )
-
-
-class FixturesDeleteButton(discord.ui.Button):
-    def __init__(self, parent_view: FixturesPanelView):
-        self.parent_view = parent_view
-        super().__init__(label="Delete Fixture", style=discord.ButtonStyle.danger)
-
-    async def callback(self, interaction: discord.Interaction):
-        fixture_id = self.parent_view.selection.fixture_id
-        if fixture_id is None:
-            await interaction.response.send_message("Select an open fixture first.", ephemeral=True)
-            return
-
-        fixture = await self.parent_view.admin_cog.db.get_fixture_by_id(fixture_id)
-        if fixture is None or fixture["status"] != "open":
-            await interaction.response.send_message(
-                "Only open fixtures can be deleted from the panel.", ephemeral=True
-            )
-            return
-
-        confirm_view = DeleteConfirmView(
-            self.parent_view.admin_cog.db,
-            self.parent_view.owner_user_id,
-            fixture_id,
-            fixture["week_number"],
-        )
-        await interaction.response.send_message(
-            f"Delete week {fixture['week_number']} and all related predictions/results/scores?",
-            view=confirm_view,
-            ephemeral=True,
-        )
-
-
-class ViewPredictionsButton(discord.ui.Button):
-    def __init__(self, parent_view: PredictionsPanelView):
-        self.parent_view = parent_view
-        super().__init__(label="View Predictions", style=discord.ButtonStyle.secondary)
-
-    async def callback(self, interaction: discord.Interaction):
-        fixture_id = self.parent_view.selection.fixture_id
-        if fixture_id is None:
-            await interaction.response.send_message("Select a fixture first.", ephemeral=True)
-            return
-
-        try:
-            (
-                fixture,
-                predictions,
-            ) = await self.parent_view.admin_cog.service.get_fixture_prediction_summary(fixture_id)
-        except ValueError as exc:
-            await interaction.response.send_message(str(exc), ephemeral=True)
-            return
-
-        lines = [f"**Week {fixture['week_number']} Predictions**"]
-        for prediction in predictions:
-            status = _prediction_status_text(prediction)
-            scores = ", ".join(prediction["predictions"])
-            lines.append(f"- {prediction['user_name']}: {scores} ({status})")
-
-        await interaction.response.send_message("\n".join(lines), ephemeral=True)
-
-
-class ReplacePredictionButton(discord.ui.Button):
-    def __init__(self, parent_view: PredictionsPanelView):
-        self.parent_view = parent_view
-        super().__init__(label="Replace Prediction", style=discord.ButtonStyle.primary)
-
-    async def callback(self, interaction: discord.Interaction):
-        fixture_id = self.parent_view.selection.fixture_id
-        user_id = self.parent_view.selection.user_id
-        if fixture_id is None or user_id is None:
-            await interaction.response.send_message(
-                "Select both fixture and user first.", ephemeral=True
-            )
-            return
-
-        fixture = await self.parent_view.admin_cog.db.get_fixture_by_id(fixture_id)
-        prediction = await self.parent_view.admin_cog.db.get_prediction(fixture_id, user_id)
-        if fixture is None or prediction is None:
-            await interaction.response.send_message(
-                "That prediction is no longer available.", ephemeral=True
-            )
-            return
-
-        modal = ReplacePredictionModal(self.parent_view, fixture, prediction)
-        await interaction.response.send_modal(modal)
-
-
-class ToggleWaiverButton(discord.ui.Button):
-    def __init__(self, parent_view: PredictionsPanelView):
-        self.parent_view = parent_view
-        super().__init__(label="Toggle Late Waiver", style=discord.ButtonStyle.success)
-
-    async def callback(self, interaction: discord.Interaction):
-        fixture_id = self.parent_view.selection.fixture_id
-        user_id = self.parent_view.selection.user_id
-        if fixture_id is None or user_id is None:
-            await interaction.response.send_message(
-                "Select both fixture and user first.", ephemeral=True
-            )
-            return
-
-        try:
-            (
-                fixture,
-                prediction,
-                recalculation,
-            ) = await self.parent_view.admin_cog.service.toggle_late_penalty_waiver(
-                fixture_id,
-                user_id,
-            )
-        except ValueError as exc:
-            await interaction.response.send_message(str(exc), ephemeral=True)
-            return
-
-        status = "enabled" if prediction["late_penalty_waived"] else "disabled"
-        self.parent_view.selection.status_message = (
-            f"Late waiver {status} for {prediction['user_name']} in week {fixture['week_number']}."
-        )
-        if recalculation is not None:
-            self.parent_view.selection.status_message += " Scores were recalculated."
-
-        await self.parent_view.load_user_options()
-        await interaction.response.edit_message(
-            content=self.parent_view.render_content(),
-            view=self.parent_view,
-        )
-
-
-class ViewResultsButton(discord.ui.Button):
-    def __init__(self, parent_view: ResultsPanelView):
-        self.parent_view = parent_view
-        super().__init__(label="View Results", style=discord.ButtonStyle.secondary)
-
-    async def callback(self, interaction: discord.Interaction):
-        fixture_id = self.parent_view.selection.fixture_id
-        if fixture_id is None:
-            await interaction.response.send_message("Select a fixture first.", ephemeral=True)
-            return
-
-        fixture = await self.parent_view.admin_cog.db.get_fixture_by_id(fixture_id)
-        results = await self.parent_view.admin_cog.db.get_results(fixture_id)
-        if fixture is None:
-            await interaction.response.send_message("Fixture not found.", ephemeral=True)
-            return
-        if not results:
-            await interaction.response.send_message(
-                "No results saved for that fixture yet.", ephemeral=True
-            )
-            return
-
-        lines = [f"**Week {fixture['week_number']} Results**"]
-        for index, (game, result) in enumerate(zip(fixture["games"], results, strict=False), 1):
-            lines.append(_format_prediction_line(index, game, result))
-        await interaction.response.send_message("\n".join(lines), ephemeral=True)
-
-
-class CorrectResultsButton(discord.ui.Button):
-    def __init__(self, parent_view: ResultsPanelView):
-        self.parent_view = parent_view
-        super().__init__(label="Correct Results", style=discord.ButtonStyle.primary)
-
-    async def callback(self, interaction: discord.Interaction):
-        fixture_id = self.parent_view.selection.fixture_id
-        if fixture_id is None:
-            await interaction.response.send_message("Select a fixture first.", ephemeral=True)
-            return
-
-        fixture = await self.parent_view.admin_cog.db.get_fixture_by_id(fixture_id)
-        if fixture is None:
-            await interaction.response.send_message("Fixture not found.", ephemeral=True)
-            return
-        if not await self.parent_view.admin_cog.db.get_results(fixture_id):
-            await interaction.response.send_message(
-                "No results are stored for that fixture yet. Use `/admin results enter` first.",
-                ephemeral=True,
-            )
-            return
-
-        modal = CorrectResultsModal(self.parent_view, fixture)
-        await interaction.response.send_modal(modal)
-
-
-class ReplacePredictionModal(discord.ui.Modal):
-    def __init__(self, parent_view: PredictionsPanelView, fixture: dict, prediction: dict):
-        super().__init__(title=f"Replace Week {fixture['week_number']} Prediction")
-        self.parent_view = parent_view
-        self.fixture = fixture
-        self.prediction = prediction
-        self.predictions_input = discord.ui.TextInput(
-            label="Predictions",
-            style=discord.TextStyle.paragraph,
-            placeholder="One line per match, e.g. Team A - Team B 2:1",
-            default="\n".join(
-                _format_prediction_line(index, game, result)
-                for index, (game, result) in enumerate(
-                    zip(fixture["games"], prediction["predictions"], strict=False),
-                    1,
-                )
-            ),
-            required=True,
-            max_length=4000,
-        )
-        self.add_item(self.predictions_input)
-
-    async def on_submit(self, interaction: discord.Interaction):
-        if not is_admin(interaction):
-            await interaction.response.send_message(
-                "You no longer have permission to use admin commands.", ephemeral=True
-            )
-            return
-
-        try:
-            (
-                fixture,
-                updated_prediction,
-                recalculation,
-            ) = await self.parent_view.admin_cog.service.replace_prediction(
-                self.fixture["id"],
-                self.prediction["user_id"],
-                self.predictions_input.value,
-                str(interaction.user.id),
-            )
-        except ValueError as exc:
-            await interaction.response.send_message(str(exc), ephemeral=True)
-            return
-
-        self.parent_view.selection.status_message = f"Replaced {updated_prediction['user_name']}'s prediction in week {fixture['week_number']}."
-        if recalculation is not None:
-            self.parent_view.selection.status_message += " Scores were recalculated."
-
-        await self.parent_view.load_user_options()
-        await interaction.response.edit_message(
-            content=self.parent_view.render_content(),
-            view=self.parent_view,
-        )
-
-
-class CorrectResultsModal(discord.ui.Modal):
-    def __init__(self, parent_view: ResultsPanelView, fixture: dict):
-        super().__init__(title=f"Correct Week {fixture['week_number']} Results")
-        self.parent_view = parent_view
-        self.fixture = fixture
-        self.results_input = discord.ui.TextInput(
-            label="Results",
-            style=discord.TextStyle.paragraph,
-            placeholder="One line per match, e.g. Team A - Team B 2:1",
-            required=True,
-            max_length=4000,
-        )
-        self.add_item(self.results_input)
-
-    async def on_submit(self, interaction: discord.Interaction):
-        if not is_admin(interaction):
-            await interaction.response.send_message(
-                "You no longer have permission to use admin commands.", ephemeral=True
-            )
-            return
-
-        try:
-            (
-                fixture,
-                _results,
-                recalculation,
-            ) = await self.parent_view.admin_cog.service.correct_results(
-                self.fixture["id"],
-                self.results_input.value,
-            )
-        except ValueError as exc:
-            await interaction.response.send_message(str(exc), ephemeral=True)
-            return
-
-        self.parent_view.selection.status_message = (
-            f"Saved corrected results for week {fixture['week_number']}."
-        )
-        if recalculation is not None:
-            self.parent_view.selection.status_message += " Scores were recalculated."
-
-        await interaction.response.edit_message(
-            content=self.parent_view.render_content(),
-            view=self.parent_view,
-        )
-
-
-class DeleteConfirmView(discord.ui.View):
-    """View for confirming fixture deletion."""
-
-    def __init__(self, db: Database, user_id: str, fixture_id: int, week_number: int):
-        super().__init__(timeout=60)
-        self.db = db
-        self.user_id = user_id
-        self.fixture_id = fixture_id
-        self.week_number = week_number
-
-    @discord.ui.button(label="Yes, Delete", style=discord.ButtonStyle.red)
-    async def confirm(self, interaction: discord.Interaction, _button: discord.ui.Button):
-        if str(interaction.user.id) != self.user_id:
-            await interaction.response.send_message(
-                "You don't have permission to do this!", ephemeral=True
-            )
-            return
-        if not is_admin(interaction):
-            await interaction.response.send_message(
-                "You no longer have permission to use admin commands.", ephemeral=True
-            )
-            return
-
-        await self.db.delete_fixture(self.fixture_id)
-        await interaction.response.edit_message(
-            content=f"**Week {self.week_number} Fixture Deleted!**",
-            view=None,
-        )
-
-    @discord.ui.button(label="Cancel", style=discord.ButtonStyle.gray)
-    async def cancel(self, interaction: discord.Interaction, _button: discord.ui.Button):
-        if str(interaction.user.id) != self.user_id:
-            await interaction.response.send_message(
-                "You don't have permission to do this!", ephemeral=True
-            )
-            return
-        if not is_admin(interaction):
-            await interaction.response.send_message(
-                "You no longer have permission to use admin commands.", ephemeral=True
-            )
-            return
-
-        await interaction.response.edit_message(
-            content="Deletion cancelled. The fixture is still active.",
-            view=None,
         )
